refactor: last changes, paddings and button styles
