--- lib/app/config/app_paddings.dart
@@ -1,6 +1,0 @@
-abstract class AppPaddings {
-  static const horizontalMedium = 20.0;
-  static const verticalMedium = 10.0;
-  static const verticalSmall = 7.0;
-  static const inner = 16.0;
-}
