+++ lib/app/config/ui_config.dart
@@ -1,5 +1,12 @@
 import "package:flutter/material.dart";
 
+abstract class AppPaddings {
+  static const horizontalMedium = 20.0;
+  static const verticalMedium = 10.0;
+  static const verticalSmall = 7.0;
+  static const inner = 16.0;
+}
+
 abstract class BottomSheetConfig {
   // fixed button
   static const fixedBottomSpace = 95.0;
