+++ lib/features/route_map/widgets/modals/end_route_modal.dart
@@ -1,6 +1,5 @@
 import "package:flutter/material.dart";
 
-import "../../../../app/config/app_paddings.dart";
 import "../../../../app/config/ui_config.dart";
 import "../../../../app/l10n/l10n.dart";
 import "../../../../app/theme/app_theme.dart";
